add: 날짜별 예약 가능 여부 관리 CRUD API 추가
