+++ common-api/src/main/java/com/reservation/commonapi/host/query/HostRoomAvailabilityQueryCondition.java
@@ -1,0 +1,10 @@
+package com.reservation.commonapi.host.query;
+
+import java.time.LocalDate;
+
+public record HostRoomAvailabilityQueryCondition(
+	Long roomTypeId,
+	LocalDate startDate,
+	LocalDate endDate
+) {
+}
