+++ common-model/src/main/java/com/reservation/commonmodel/accommodation/RoomAvailabilityDto.java
@@ -1,0 +1,11 @@
+package com.reservation.commonmodel.accommodation;
+
+import java.time.LocalDate;
+
+public record RoomAvailabilityDto(
+	Long id,
+	Long roomTypeId,
+	LocalDate date,
+	Integer availableCount // 예약 가능 개수
+) {
+}
