+++ host/src/main/java/com/reservation/host/accommodation/availability/controller/dto/request/CreateRoomAvailabilityRequest.java
@@ -1,0 +1,10 @@
+package com.reservation.host.accommodation.availability.controller.dto.request;
+
+import java.time.LocalDate;
+
+public record CreateRoomAvailabilityRequest(
+	Long roomTypeId,
+	LocalDate date,
+	int availableCount
+) {
+}
