+++ host/src/main/java/com/reservation/host/accommodation/availability/controller/dto/request/RoomAvailabilitySearchCondition.java
@@ -1,0 +1,10 @@
+package com.reservation.host.accommodation.availability.controller.dto.request;
+
+import java.time.LocalDate;
+
+public record RoomAvailabilitySearchCondition(
+	Long roomTypeId,
+	LocalDate startDate,
+	LocalDate endDate
+) {
+}
