+++ host/src/main/java/com/reservation/host/accommodation/availability/controller/dto/request/UpdateRoomAvailabilityRequest.java
@@ -1,0 +1,11 @@
+package com.reservation.host.accommodation.availability.controller.dto.request;
+
+import java.time.LocalDate;
+
+public record UpdateRoomAvailabilityRequest(
+	Long id,
+	Long roomTypeId,
+	LocalDate date,
+	int availableCount
+) {
+}
