--- host/src/main/java/com/reservation/host/accommodation/service/mapper/RoomImageDtoMapper.java
@@ -1,17 +1,0 @@
-package com.reservation.host.accommodation.service.mapper;
-
-import com.reservation.commonmodel.accommodation.RoomImageDto;
-import com.reservation.host.accommodation.controller.dto.request.UpdateRoomImagesRequest;
-
-public class RoomImageDtoMapper {
-	public static RoomImageDto fromUpdateRoomImage(UpdateRoomImagesRequest.UpdateRoomImage roomImageDto,
-		Long roomTypeId, String uploadUrl) {
-		return new RoomImageDto(
-			roomImageDto.id(),
-			roomTypeId,
-			uploadUrl,
-			roomImageDto.displayOrder(),
-			roomImageDto.isMainImage()
-		);
-	}
-}
